pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Carlos Chiarelli <ca.chiarelli.97@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
motor = "^3.6.0"
httpx = "^0.27.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
email-validator = "^2.3.0"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
alembic = "^1.13.1"
ipython = "^8.22.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
pytest-sugar = "^1.0.0"
pytest-clarity = "^1.0.1"
pytest-pretty = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
